win32 port : warnings removed

+++ pvDataApp/factory/PVUnion.cpp
@@ -130,7 +130,7 @@ void PVUnion::set(int32 index, PVFieldPtr const & value)
 
 void PVUnion::set(String const & fieldName, PVFieldPtr const & value)
 {
-    int32 index = variant ? -1 : unionPtr->getFieldIndex(fieldName);
+    int32 index = variant ? -1 : static_cast<int32>(unionPtr->getFieldIndex(fieldName));
     if (index == -1)
         throw std::invalid_argument("no such fieldName");
